refactor: extract profile selection
